修复相乘溢出：RayDirections resize 参数类型更正
将 Camera::RecalculateRayDirections() 中 m_RayDirections 的 resize 参数改为先将 m_ViewportWidth 和 m_ViewportHeight 转为 size_t 后再相乘，避免 uint32_t 相乘时的溢出风险，提高代码健壮性。

+++ RayTracing/src/Camera.cpp
@@ -137,7 +137,9 @@ void Camera::RecalculateView()
 
 void Camera::RecalculateRayDirections()
 {
-	m_RayDirections.resize(m_ViewportWidth * m_ViewportHeight);
+	m_RayDirections.resize(
+		static_cast<size_t>(m_ViewportWidth) * static_cast<size_t>(m_ViewportHeight)
+	);
 
 	for (uint32_t y = 0; y < m_ViewportHeight; y++)
 	{
